+++ modules/network/src/test/java/org/apache/ignite/internal/network/AllTypesMessageGenerator.java
@@ -25,13 +25,13 @@ import java.util.BitSet;
 import java.util.Random;
 import java.util.UUID;
 import java.util.stream.IntStream;
+import org.apache.ignite.internal.network.messages.AllTypesMessage;
+import org.apache.ignite.internal.network.messages.AllTypesMessageBuilder;
+import org.apache.ignite.internal.network.messages.AllTypesMessageImpl;
+import org.apache.ignite.internal.network.messages.TestMessagesFactory;
 import org.apache.ignite.lang.IgniteUuid;
 import org.apache.ignite.lang.IgniteUuidGenerator;
 import org.apache.ignite.network.NetworkMessage;
-import org.apache.ignite.network.messages.AllTypesMessage;
-import org.apache.ignite.network.messages.AllTypesMessageBuilder;
-import org.apache.ignite.network.messages.AllTypesMessageImpl;
-import org.apache.ignite.network.messages.TestMessagesFactory;
 import org.jetbrains.annotations.Nullable;
 
 /**
