+++ modules/network/src/test/java/org/apache/ignite/internal/network/messages/NetworkMessage2.java
@@ -16,7 +16,7 @@
  */
 
 
-package org.apache.ignite.network.messages;
+package org.apache.ignite.internal.network.messages;
 
 import java.io.Serializable;
 import org.apache.ignite.network.NetworkMessage;
