+++ modules/network/src/test/java/org/apache/ignite/internal/network/messages/TestMessage.java
@@ -15,29 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.network.messages;
+package org.apache.ignite.internal.network.messages;
 
-import org.apache.ignite.network.annotations.MessageGroup;
+import java.io.Serializable;
+import java.util.Map;
+import org.apache.ignite.network.NetworkMessage;
+import org.apache.ignite.network.annotations.Transferable;
 
 /**
- * Test message types.
+ * {@link NetworkMessage} implementation.
  */
-@MessageGroup(groupName = "TestMessages", groupType = 2)
-public class TestMessageTypes {
-    /** Type of {@link SerializationOrderMessage}. */
-    public static final short SERIALIZATION_ORDER_MESSAGE = 1;
+@Transferable(TestMessageTypes.TEST)
+public interface TestMessage extends NetworkMessage, Serializable {
+    String msg();
 
-    /** Type of {@link InheritedMessage}. */
-    public static final short INHERITED_MESSAGE = 2;
-
-    /** Type of {@link EmptyMessage}. */
-    public static final short EMPTY_MESSAGE = 3;
-
-    public static final short ALL_TYPES = 4;
-
-    public static final short TEST = 5;
-
-    public static final short NESTED_MESSAGE = 6;
-
-    public static final short TEST_MESSAGE_WITH_MARSHALLABLE = 7;
+    Map<Integer, String> map();
 }
