+++ modules/network/src/test/java/org/apache/ignite/internal/network/netty/InboundDecoderTest.java
@@ -36,14 +36,14 @@ import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 import org.apache.ignite.internal.network.AllTypesMessageGenerator;
 import org.apache.ignite.internal.network.direct.DirectMessageWriter;
+import org.apache.ignite.internal.network.messages.AllTypesMessage;
+import org.apache.ignite.internal.network.messages.NestedMessageMessage;
+import org.apache.ignite.internal.network.messages.TestMessage;
+import org.apache.ignite.internal.network.messages.TestMessagesFactory;
 import org.apache.ignite.internal.network.serialization.PerSessionSerializationService;
 import org.apache.ignite.internal.network.serialization.SerializationService;
 import org.apache.ignite.internal.network.serialization.UserObjectSerializationContext;
 import org.apache.ignite.network.NetworkMessage;
-import org.apache.ignite.network.messages.AllTypesMessage;
-import org.apache.ignite.network.messages.NestedMessageMessage;
-import org.apache.ignite.network.messages.TestMessage;
-import org.apache.ignite.network.messages.TestMessagesFactory;
 import org.apache.ignite.network.serialization.MessageSerializationRegistry;
 import org.apache.ignite.network.serialization.MessageSerializer;
 import org.apache.ignite.network.serialization.TestMessageSerializationRegistryImpl;
