+++ modules/network/src/test/java/org/apache/ignite/internal/network/processor/EmptyMessageTest.java
@@ -20,9 +20,9 @@ package org.apache.ignite.internal.network.processor;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import org.apache.ignite.network.messages.EmptyMessage;
-import org.apache.ignite.network.messages.EmptyMessageImpl;
-import org.apache.ignite.network.messages.TestMessagesFactory;
+import org.apache.ignite.internal.network.messages.EmptyMessage;
+import org.apache.ignite.internal.network.messages.EmptyMessageImpl;
+import org.apache.ignite.internal.network.messages.TestMessagesFactory;
 import org.junit.jupiter.api.Test;
 
 /**
